+++ packages/maskbook/src/utils/theme.ts
@@ -1,7 +1,8 @@
 import { ValueRef } from '@dimensiondev/holoflows-kit'
 import { useValueRef, or } from '@masknet/shared'
 import { safeUnreachable } from '@dimensiondev/kit'
-import { Appearance, LanguageOptions, SupportedLanguages } from '@masknet/theme'
+import { Appearance } from '@masknet/theme'
+import { LanguageOptions, SupportedLanguages } from '@masknet/public-api'
 import { unstable_createMuiStrictModeTheme, useMediaQuery } from '@material-ui/core'
 import { blue, green, grey, orange, red } from '@material-ui/core/colors'
 import { jaJP, koKR, zhTW, zhCN, esES, itIT, ruRU, faIR, frFR } from '@material-ui/core/locale/index'
