+++ packages/public-api/src/web.ts
@@ -55,3 +55,31 @@ export enum TradeProvider {
     DODO = 7,
     UNISWAP_V3 = 8,
 }
+/** Supported language settings */
+export enum LanguageOptions {
+    __auto__ = 'auto',
+    enUS = 'en-US',
+    zhCN = 'zh-CN',
+    zhTW = 'zh-TW',
+    koKR = 'ko-KR',
+    jaJP = 'ja-JP',
+    esES = 'es-ES',
+    faIR = 'fa-IR',
+    itIT = 'it-IT',
+    ruRU = 'ru-RU',
+    frFR = 'fr-FR',
+}
+
+/** Supported display languages */
+export enum SupportedLanguages {
+    enUS = 'en-US',
+    zhCN = 'zh-CN',
+    zhTW = 'zh-TW',
+    koKR = 'ko-KR',
+    jaJP = 'ja-JP',
+    esES = 'es-ES',
+    faIR = 'fa-IR',
+    itIT = 'it-IT',
+    ruRU = 'ru-RU',
+    frFR = 'fr-FR',
+}
