+++ packages/theme/src/theme.ts
@@ -54,31 +54,3 @@ export enum Appearance {
     light = 'light',
     dark = 'dark',
 }
-
-/** Supported display languages */
-export enum SupportedLanguages {
-    enUS = 'en-US',
-    zhCN = 'zh-CN',
-    zhTW = 'zh-TW',
-    koKR = 'ko-KR',
-    jaJP = 'ja-JP',
-    esES = 'es-ES',
-    faIR = 'fa-IR',
-    itIT = 'it-IT',
-    ruRU = 'ru-RU',
-    frFR = 'fr-FR',
-}
-/** Supported language settings */
-export enum LanguageOptions {
-    __auto__ = 'auto',
-    enUS = 'en-US',
-    zhCN = 'zh-CN',
-    zhTW = 'zh-TW',
-    koKR = 'ko-KR',
-    jaJP = 'ja-JP',
-    esES = 'es-ES',
-    faIR = 'fa-IR',
-    itIT = 'it-IT',
-    ruRU = 'ru-RU',
-    frFR = 'fr-FR',
-}
